+++ proj/src/Point2D.cpp
@@ -46,8 +46,8 @@ bool Point2D::operator!=(const Point2D& other)const{
     return !(this->x == other.x && this->y == other.y);
 }
 
-int Point2D::getX()const{return x;}
-int Point2D::getY()const{return y;}
+double Point2D::getX()const{return x;}
+double Point2D::getY()const{return y;}
 void Point2D::setX(const double _x){x=_x;}
 void Point2D::setY(const double _y){y=_y;}
 void Point2D::setPoint(const double _x, const double _y){x=_x;y=_y;}
